--- app/src/main/java/de/projektss17/bonpix/fragments/F_Laeden.java
@@ -1,64 +1,0 @@
-package de.projektss17.bonpix.fragments;
-
-
-import android.os.Bundle;
-import android.app.Fragment;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.DividerItemDecoration;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import de.projektss17.bonpix.R;
-import de.projektss17.bonpix.S;
-import de.projektss17.bonpix.daten.C_Bon;
-import de.projektss17.bonpix.daten.C_Laden;
-import de.projektss17.bonpix.daten.C_Laeden_Adapter;
-
-
-
-
-/**
- * A fragment representing a list of Items.
- * <p/>
- * Activities containing this fragment MUST implement the {@link OnListFragmentInteractionListener}
- * interface.
- */
-public class F_Laeden extends Fragment {
-    private List<C_Bon> bonsList = new ArrayList<>();
-    private List<C_Laden> shopList = new ArrayList<>();
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-
-        View rootView = inflater.inflate(R.layout.box_laeden_screen, container, false);
-
-        RecyclerView recyclerViewLaeden = (RecyclerView) rootView.findViewById(R.id.laeden_view);
-        C_Laeden_Adapter mAdapter = new C_Laeden_Adapter(shopList);
-        prepareBonData();
-        RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(container.getContext());
-        recyclerViewLaeden.setLayoutManager(mLayoutManager);
-        recyclerViewLaeden.addItemDecoration(
-                new DividerItemDecoration(getActivity(), DividerItemDecoration.VERTICAL));
-        recyclerViewLaeden.setItemAnimator(new DefaultItemAnimator());
-        recyclerViewLaeden.setAdapter(mAdapter);
-        mAdapter.notifyDataSetChanged();
-        return rootView;
-    }
-
-    /**
-     * Set Data for RecyclerView Bons
-     */
-    private void prepareBonData(){
-        shopList.clear();
-        for(C_Laden shop : S.dbHandler.getAllLaeden(S.db)){
-            this.shopList.add(shop);
-        }
-    }
-}
