+++ app/src/main/res/layout/box_laeden_content.xml
@@ -5,6 +5,10 @@
     android:paddingTop="@dimen/title_bar"
     android:orientation="vertical">
 
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/view_laeden"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
 
 </RelativeLayout>
