+++ app/src/main/res/layout/box_laeden_screen.xml
@@ -11,44 +11,25 @@
         android:layout_height="wrap_content"
         android:theme="@style/AppTheme.AppBarOverlay">
 
-        <RelativeLayout
+        <android.support.v7.widget.Toolbar
+            android:id="@+id/toolbar"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_weight="1">
-
-            <android.support.v7.widget.Toolbar
-                android:id="@+id/toolbar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:background="?attr/colorPrimary"
-                app:popupTheme="@style/AppTheme.PopupOverlay" />
-
-        </RelativeLayout>
+            android:layout_height="?attr/actionBarSize"
+            android:background="?attr/colorPrimary"
+            app:popupTheme="@style/AppTheme.PopupOverlay" />
 
     </android.support.design.widget.AppBarLayout>
 
-    <RelativeLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="55dp"
-        android:id="@+id/relativeLayout">
+    <include layout="@layout/box_laeden_content" />
 
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/laeden_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    </RelativeLayout>
 
     <android.support.design.widget.FloatingActionButton
-        android:id="@+id/laeden_fab"
+        android:id="@+id/laeden_fap"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="top|left"
+        android:layout_gravity="bottom|end"
         android:layout_margin="@dimen/fab_margin"
-        app:layout_anchor="@+id/relativeLayout"
-        app:layout_anchorGravity="bottom|right"
-        app:srcCompat="@mipmap/ic_playlist_add_white_24dp" />
-
+        android:gravity="center_vertical"
+        app:srcCompat="@mipmap/ic_add_white_24dp" />
 
 </android.support.design.widget.CoordinatorLayout>
